Refactored error handling in actions
name fixes
