+++ src/reducers/ShowOfferReducer.js
@@ -2,7 +2,7 @@ import {
   GET_OFFER_BY_ID,
   GET_OFFER_BY_ID_FAILURE,
   GET_OFFER_BY_ID_SUCCESS
-} from "../actions/ShowOfferAction";
+} from "../actions/ShowOfferActions";
 
 const initialState = {
   isFetching: false,
